+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Check for more Steam interface versions
 * [ImGui] Show redraw fps on main menu
 * [ImGui] Idling when no activity
+* [ImGui] New audio debug window
 
 ### Changed